Saved space in tables

+++ data-loading.sql
@@ -4,14 +4,14 @@ SET search_path TO f23_group2;
 --Episodes
 DROP TABLE IF EXISTS office_episodes CASCADE;
 CREATE TABLE office_episodes (
-    season INT,
-    episode_in_season INT,
-    episode_overall INT,
+    season NUMERIC(1),
+    episode_in_season NUMERIC(2),
+    episode_overall NUMERIC(3),
     title TEXT,
     director TEXT,
     writer TEXT,
     air_date DATE,
-    prod_code INT,
+    prod_code NUMERIC(4),
     us_viewers FLOAT
 );
 ALTER TABLE office_episodes ADD PRIMARY KEY(prod_code); 
@@ -21,12 +21,12 @@ ALTER TABLE f23_group2.office_episodes OWNER TO f23_group2;
 --IMDB ratings
 DROP TABLE IF EXISTS office_ratings CASCADE;
 CREATE TABLE office_ratings (
-   season INT,
-   episode_in_season INT, 
+   season NUMERIC(1),
+   episode_in_season NUMERIC(2), 
    title TEXT,
    air_date DATE,
-   rating FLOAT,
-   total_votes INT,
+   rating NUMERIC(2, 1),
+   total_votes NUMERIC(5),
    description TEXT
 );
 ALTER TABLE office_ratings ADD PRIMARY KEY(season, episode_in_season); 
@@ -40,9 +40,9 @@ ALTER TABLE f23_group2.office_ratings OWNER TO f23_group2;
 --Then parks and rec
 DROP TABLE IF EXISTS parks_and_rec_episodes CASCADE;
 CREATE TABLE parks_and_rec_episodes (
-    season INT,
-    episode_num_in_season INT,
-    episode_num_overall INT,
+    season NUMERIC(1),
+    episode_num_in_season NUMERIC(2),
+    episode_num_overall NUMERIC(3),
     title TEXT,
     directed_by TEXT,
     writen_by TEXT,
@@ -56,12 +56,12 @@ ALTER TABLE f23_group2.parks_and_rec_episodes OWNER TO f23_group2;
 --IMDB ratings
 DROP TABLE IF EXISTS parks_and_rec_ratings CASCADE;
 CREATE TABLE parks_and_rec_ratings (
-   season INT,
-   episode_in_season INT, 
+   season NUMERIC(1),
+   episode_in_season NUMERIC(2), 
    title TEXT,
    air_date DATE,
-   rating FLOAT,
-   total_votes INT,
+   rating NUMERIC(2, 1),
+   total_votes NUMERIC(5),
    description TEXT
 );
 ALTER TABLE parks_and_rec_ratings ADD PRIMARY KEY(season, episode_in_season);
